1. 开始修正功能分析模块，为自动化报告做准备

+++ src/main/java/com/novelbio/analysis/seq/mapping/MapTophat.java
@@ -273,7 +273,7 @@ public class MapTophat implements MapRNA{
 	}
 	private String getIndelLen() {
 		String param = "--max-insertion-length " + indelLen + " --max-deletion-length " + 
-				indelLen + " " + "--read-gap-length " + (int)(indelLen * 1.2) + " --read-edit-dist " + (indelLen*1.5 + mismatch) + " ";
+				indelLen + " " + "--read-gap-length " + (int)(indelLen * 1.2) + " --read-edit-dist " + (int)(indelLen*1.5 + mismatch) + " ";
 		return param;
 	}
